src/requirements.txt
fastapi==0.104.1
requests==2.31.0
pyyaml==6.0.1
colorama
openai==1.3.7
loguru
uvicorn==0.24.0
schedule
python-dotenv
parameterized==0.8.1
Locust